Turned Profile Edit into two pages, one for registered users and one for unregistered users
