+++ client/src/pages/Profile/Profile.jsx
@@ -821,12 +821,14 @@ const ProfilePageHeader = ({ leader, editButton }) => {
               </>
             )}
           </div>
-          {editButton !== false && !isRegistered ? (
+          {editButton !== false && isRegistered ? (
             <Link to={'/profile-edit'} className={'profile-edit-icon-link no-link-style'}>
               <img src={EditIcon} alt={'edit'} className={'profile-edit-icon'} />
             </Link>
           ) : (
-            <></>
+            <Link to={'/profile-edit-unregistered'} className={'profile-edit-icon-link no-link-style'}>
+              <img src={EditIcon} alt={'edit'} className={'profile-edit-icon'} />
+            </Link>
           )}
         </div>
       </div>
