+++ client/src/pages/ProfileEditUnregistered/ProfileEditUnregistered.scss
@@ -1,0 +1,144 @@
+@import '../../scssStyles/mixins';
+
+.profile-edit-page {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  transition: transform 700ms, opacity 700ms;
+}
+.profile-edit-logo {
+  width: 140px;
+  height: 140px;
+  margin-bottom: 10px;
+  transition: transform 500ms;
+}
+
+.profile-edit-logo-expand {
+  transform: scale(0.6);
+}
+
+.profile-edit-container {
+  background-color: var(--container-bg);
+  border-radius: 15px;
+  padding: 40px min(150px, 5%);
+  margin: 50px;
+  box-shadow: var(--shadow);
+  width: 95vw;
+  max-width: 1000px;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  h1 {
+    text-align: center;
+    font-size: 40px;
+    width: 100%;
+  }
+  h3 {
+    font-size: 17px;
+    text-align: center;
+    margin-bottom: 15px;
+    width: 100%;
+  }
+}
+
+.profile-edit-input {
+  width: 100%;
+}
+
+.profile-edit-button {
+  margin-top: 20px;
+}
+
+.profile-edit-page-disappear {
+  transform: translateY(-35vh);
+  opacity: 0;
+  transition: transform 1200ms, opacity 1200ms;
+  user-select: none;
+  pointer-events: none;
+}
+
+.profile-edit-loading {
+  z-index: -1;
+  position: fixed;
+  top: calc(50vh - 30px);
+  left: calc(50vw - 30px);
+  transition: transform 400ms, opacity 400ms;
+  opacity: 0;
+  transform: scale(0.5);
+}
+
+.profile-edit-loading-appear {
+  transition: transform 400ms, opacity 400ms;
+  opacity: 1;
+  transform: scale(1);
+}
+
+.profile-edit-loading-disappear {
+  transition: transform 1500ms, opacity 900ms;
+  opacity: 0;
+  transform: scale(10);
+}
+
+.profile-edit-success {
+  color: var(--text-dynamic);
+  position: fixed;
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  // flex-direction: column;
+  // align-items: center;
+  // justify-content: center;
+  user-select: none;
+  pointer-events: none;
+
+  padding: 0px 25px;
+  text-align: center;
+  @include devices(tablet) {
+    padding: 0px 12px;
+  }
+  h1 {
+    font-size: 50px;
+    margin-bottom: 20px;
+    @include devices(tablet) {
+      font-size: 35px;
+    }
+  }
+  h2 {
+    font-size: 35px;
+    margin-bottom: 19px;
+    @include devices(tablet) {
+      font-size: 30px;
+    }
+  }
+  h3 {
+    font-size: 25px;
+    margin-bottom: 19px;
+    @include devices(tablet) {
+      font-size: 20px;
+    }
+  }
+  @for $i from 2 to 6 {
+    div :nth-child(#{$i}) {
+      opacity: 0;
+      transform: scale(0.5);
+      transition: transform 1000ms ease-in-out ($i - 1.5) * 700ms,
+        opacity 1000ms ease-in-out ($i - 1.5) * 700ms;
+    }
+  }
+}
+
+.profile-edit-success-appear {
+  position: unset;
+
+  user-select: unset;
+  pointer-events: unset;
+  @for $i from 2 to 6 {
+    div :nth-child(#{$i}) {
+      opacity: 1;
+      transform: scale(1);
+      transition: transform 1000ms ease-in-out ($i - 1.5) * 700ms,
+        opacity 1000ms ease-in-out ($i - 1.5) * 700ms;
+    }
+  }
+}
